fix(player): resolve pausing when gesture is active and move slider positions
