+++ feature/mobile/player/src/main/res/drawable/brightness_half.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="960"
+    android:viewportHeight="960">
+  <path
+      android:pathData="M354.46,780L252.31,780q-29.92,0 -51.12,-21.19Q180,737.61 180,707.69v-102.15l-74.31,-74.93q-10.23,-10.84 -15.46,-23.99Q85,493.46 85,480t5.23,-26.62q5.23,-13.15 15.46,-23.99L180,354.46v-102.15q0,-29.92 21.19,-51.12Q222.39,180 252.31,180h102.15l74.93,-74.31q10.84,-10.23 23.99,-15.46Q466.54,85 480,85t26.62,5.23q13.15,5.23 23.99,15.46L605.54,180h102.15q29.92,0 51.12,21.19Q780,222.39 780,252.31v102.15l74.31,74.93q10.23,10.84 15.46,23.99Q875,466.54 875,480t-5.23,26.62q-5.23,13.15 -15.46,23.99L780,605.54v102.15q0,29.92 -21.19,51.12Q737.61,780 707.69,780L605.54,780l-74.93,74.31q-10.84,10.23 -23.99,15.46Q493.46,875 480,875t-26.62,-5.23q-13.15,-5.23 -23.99,-15.46L354.46,780ZM380,720l91.15,91.15q3.46,3.46 8.85,3.46t8.85,-3.46L580,720h127.69q5.39,0 8.85,-3.46t3.46,-8.85L720,580l91.15,-91.15q3.46,-3.46 3.46,-8.85t-3.46,-8.85L720,380v-127.69q0,-5.39 -3.46,-8.85t-8.85,-3.46L580,240l-91.15,-91.15q-3.46,-3.46 -8.85,-3.46t-8.85,3.46L380,240L252.31,240q-5.39,0 -8.85,3.46t-3.46,8.85L240,380l-91.15,91.15q-3.46,3.46 -3.46,8.85t3.46,8.85L240,580v127.69q0,5.39 3.46,8.85t8.85,3.46L380,720ZM480,664.61q76.46,0 130.54,-54.07 54.07,-54.08 54.07,-130.54t-54.07,-130.54Q556.46,295.39 480,295.39v369.22Z"
+      android:fillColor="#e3e3e3"/>
+</vector>
