--- feature/mobile/player/src/main/res/drawable/round_wb_sunny_24.xml
@@ -1,5 +1,0 @@
-<vector android:height="24dp" android:tint="#000000"
-    android:viewportHeight="24" android:viewportWidth="24"
-    android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
-    <path android:fillColor="@android:color/white" android:pathData="M6.05,4.14l-0.39,-0.39c-0.39,-0.39 -1.02,-0.38 -1.4,0l-0.01,0.01c-0.39,0.39 -0.39,1.02 0,1.4l0.39,0.39c0.39,0.39 1.01,0.39 1.4,0l0.01,-0.01c0.39,-0.38 0.39,-1.02 0,-1.4zM3.01,10.5L1.99,10.5c-0.55,0 -0.99,0.44 -0.99,0.99v0.01c0,0.55 0.44,0.99 0.99,0.99L3,12.49c0.56,0.01 1,-0.43 1,-0.98v-0.01c0,-0.56 -0.44,-1 -0.99,-1zM12.01,0.55L12,0.55c-0.56,0 -1,0.44 -1,0.99v0.96c0,0.55 0.44,0.99 0.99,0.99L12,3.49c0.56,0.01 1,-0.43 1,-0.98v-0.97c0,-0.55 -0.44,-0.99 -0.99,-0.99zM19.75,3.76c-0.39,-0.39 -1.02,-0.39 -1.41,-0.01l-0.39,0.39c-0.39,0.39 -0.39,1.02 0,1.4l0.01,0.01c0.39,0.39 1.02,0.39 1.4,0l0.39,-0.39c0.39,-0.39 0.39,-1.01 0,-1.4zM17.94,18.86l0.39,0.39c0.39,0.39 1.02,0.39 1.41,0 0.39,-0.39 0.39,-1.02 0,-1.41l-0.39,-0.39c-0.39,-0.39 -1.02,-0.38 -1.4,0 -0.4,0.4 -0.4,1.02 -0.01,1.41zM20,11.49v0.01c0,0.55 0.44,0.99 0.99,0.99L22,12.49c0.55,0 0.99,-0.44 0.99,-0.99v-0.01c0,-0.55 -0.44,-0.99 -0.99,-0.99h-1.01c-0.55,0 -0.99,0.44 -0.99,0.99zM12,5.5c-3.31,0 -6,2.69 -6,6s2.69,6 6,6 6,-2.69 6,-6 -2.69,-6 -6,-6zM11.99,22.45L12,22.45c0.55,0 0.99,-0.44 0.99,-0.99v-0.96c0,-0.55 -0.44,-0.99 -0.99,-0.99h-0.01c-0.55,0 -0.99,0.44 -0.99,0.99v0.96c0,0.55 0.44,0.99 0.99,0.99zM4.25,19.24c0.39,0.39 1.02,0.39 1.41,0l0.39,-0.39c0.39,-0.39 0.38,-1.02 0,-1.4l-0.01,-0.01c-0.39,-0.39 -1.02,-0.39 -1.41,0l-0.39,0.39c-0.38,0.4 -0.38,1.02 0.01,1.41z"/>
-</vector>
